Prevent extra carousel animation triggers
When manually scrolling a carousel, we want to "reset" the timer so that
we don't get an extra scroll if the timer goes off just after we've
scrolled. We do this by tracking a nonce for each carousel and sending
it with each carousel update. When messages with old nonces come in,
they can safely be ignored as they were sent before the most recent
update.